+++ PlanBuild/PlanBuildPlugin.cs
@@ -13,7 +13,7 @@ namespace PlanBuild
     {
         public const string PluginGUID = "augusdogus.Buildheim";
         public const string PluginName = "Buildheim";
-        public const string PluginVersion = "0.20.1";
+        public const string PluginVersion = "0.20.2";
 
         private ClientPlanner planner;